web/requirements.txt
Django==1.9
gunicorn==19.3.0
psycopg2==2.6.2
redis==2.10.3
dj-database-url
coverage
nose
django-debug-toolbar
django-configurations